When running ksp allow the incremental flag to not be set and still get the custom processors from config. Use testKsp to test that behavior.
